+++ app/constants/lists.js
@@ -271,13 +271,7 @@ angular.module("orsApp").constant("lists", {
       },
       fords: {
         name: "fords",
-        subgroups: [
-          "Car",
-          "Bicycle",
-          "HeavyVehicle",
-          "Wheelchair",
-          "Pedestrian"
-        ]
+        subgroups: ["Bicycle", "Wheelchair", "Pedestrian"]
       },
       steps: {
         name: "steps",